# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Flash-Archi Public API
  description: >
    API publique de Flash-Archi (pré-production) — projets, jobs,
    artefacts, versions et webhooks.

    Clés d'accès :
    - **API publique** (`/api/public/v1/**`) : clé Bearer étendue avec scopes
      `public:*`, résolue par tenant (plan : free / pro / enterprise).
    - **API interne** (`/api/flash-archi/projects/**/versions/**`)
      : clé applicative `API_KEY` (environnement), scopes internes.

    Tous les endpoints honorent le rate limiting (headers `X-RateLimit-*`).
  version: 3.3.0
  contact:
    name: Flash-Archi
servers:
- url: /
  description: "Racine \u2014 chemins complets (API publique + interne)"
...
paths:
...
  /api/public/v1/analytics/export:
    get:
      tags:
      - Analytics
      summary: Export analytics data
      operationId: exportAnalytics
      security:
      - PublicKeyAuth:
        - public:read:analytics
      parameters:
      - name: format
        in: query
        required: false
        schema:
          type: string
          enum:
          - json
          - csv
          default: json
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Analytics data export
          content:
            application/json: {}
            text/csv: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/public/v1/analytics/config:
    get:
      tags:
      - Analytics
      summary: Get analytics configuration for the tenant
      operationId: getAnalyticsConfig
      security:
      - PublicKeyAuth:
        - public:read:analytics
      responses:
        '200':
          description: Analytics configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      tags:
      - Analytics
      summary: Update analytics configuration for the tenant
      operationId: patchAnalyticsConfig
      security:
      - PublicKeyAuth:
        - public:write:analytics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsConfigInput'
      responses:
        '200':
          description: Updated analytics configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/public/v1/admin/analytics/export:
    get:
      tags:
      - Analytics
      summary: Export analytics data for all tenants (admin)
      operationId: exportAdminAnalytics
      security:
      - PublicKeyAuth:
        - admin:analytics
      parameters:
      - name: tenant_id
        in: query
        required: false
        schema:
          type: integer
      - name: format
        in: query
        required: false
        schema:
          type: string
          enum:
          - json
          - csv
          default: json
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Analytics data export
          content:
            application/json: {}
            text/csv: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  securitySchemes:
    PublicKeyAuth:
      type: http
      scheme: bearer
      description: "Cl\xE9 API publique (scope `public:*`) r\xE9solue par tenant.
        \ Le scope requis est indiqu\xE9 par endpoint. Plans : free=100 req/min, pro=500
        \ req/min, enterprise=1000 req/min.\n"
    InternalKeyAuth:
      type: http
      scheme: bearer
      description: "Cl\xE9 applicative interne `API_KEY` (routes versions)."
...
  schemas:
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AnalyticsConfig:
      type: object
      properties:
        tenant_id:
          type: integer
        cost_threshold_usd:
          type: number
          format: float
        error_rate_threshold:
          type: number
          format: float
        job_depth_threshold:
          type: integer
        queue_depth_threshold:
          type: integer
        alert_email_enabled:
          type: boolean
        alert_webhook_enabled:
          type: boolean
      required:
      - tenant_id
    AnalyticsConfigInput:
      type: object
      properties:
        cost_threshold_usd:
          type: number
          format: float
        error_rate_threshold:
          type: number
          format: float
        job_depth_threshold:
          type: integer
        queue_depth_threshold:
          type: integer
        alert_email_enabled:
          type: boolean
        alert_webhook_enabled:
          type: boolean
  responses:
    BadRequest:
      description: "Requ\xEAte invalide"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: "Authentification \xE9chou\xE9e ou cl\xE9 manquante"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Scope insuffisant pour cette ressource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: "Quota d\xE9pass\xE9"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: "Quota de la fen\xEAtre"
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: "Requ\xEAtes restantes dans la fen\xEAtre"
        X-RateLimit-Reset:
          schema:
            type: integer
          description: "Timestamp (ms) de r\xE9initialisation"
        Retry-After:
          schema:
            type: integer
          description: "Secondes avant retry autoris\xE9"